requirements.txt
# Core dependencies
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=9.0.0

# Image processing and metrics
scikit-image>=0.19.0
scikit-learn>=1.0.0
opencv-python>=4.5.0

# Perceptual loss (LPIPS)
torch>=1.10.0
lpips>=0.1.4

# Visualization
matplotlib>=3.5.0

# Utilities
tqdm>=4.62.0
